+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 0.0.4
 
-- Downgrade initl to 0.18.1 package.
+- Remove static from classes.
 - Flutter Version: 3.16.9